requirements.txt
discord.py==1.7.3
openai==0.19.0
python-dotenv
dnspython

